backend/requirements.txt
#Dependencies already installed
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
python-dotenv
